pyproject.toml
[project]
name = "openai-agent"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]",
    "openai",
    "pydantic",
    "python-dotenv",
    "uvicorn",
    "pyyaml",
]
